+++ src/argilla/server/contexts/datasets.py
@@ -42,7 +42,9 @@ from argilla.server.security.model import User
 if TYPE_CHECKING:
     from sqlalchemy.ext.asyncio import AsyncSession
 
+    from argilla.server.schemas.v1.datasets import DatasetUpdate
     from argilla.server.schemas.v1.fields import FieldUpdate
+    from argilla.server.schemas.v1.questions import QuestionUpdate
     from argilla.server.schemas.v1.suggestions import SuggestionCreate
 
 LIST_RECORDS_LIMIT = 20
@@ -125,8 +127,8 @@ async def delete_dataset(db: "AsyncSession", search_engine: SearchEngine, datase
     return dataset
 
 
-async def update_dataset(db: "AsyncSession", dataset: Dataset, dataset_update: DatasetCreate) -> Dataset:
-    params = dataset_update.dict(exclude_unset=True, exclude_none=True)
+async def update_dataset(db: "AsyncSession", dataset: Dataset, dataset_update: "DatasetUpdate") -> Dataset:
+    params = dataset_update.dict(exclude_unset=True)
     return await dataset.update(db, **params)
 
 
@@ -155,7 +157,7 @@ async def create_field(db: "AsyncSession", dataset: Dataset, field_create: Field
 
 
 async def update_field(db: "AsyncSession", field: Field, field_update: "FieldUpdate") -> Field:
-    params = field_update.dict(exclude_unset=True, exclude_none=True)
+    params = field_update.dict(exclude_unset=True)
     return await field.update(db, **params)
 
 
@@ -189,6 +191,11 @@ async def create_question(db: "AsyncSession", dataset: Dataset, question_create:
         settings=question_create.settings.dict(),
         dataset_id=dataset.id,
     )
+
+
+async def update_question(db: "AsyncSession", question: Question, question_update: "QuestionUpdate") -> Question:
+    params = question_update.dict(exclude_unset=True)
+    return await question.update(db, **params)
 
 
 async def delete_question(db: "AsyncSession", question: Question) -> Question: