+++ src/argilla/server/services/datasets.py
@@ -26,7 +26,7 @@ from argilla.server.database import get_async_db
 from argilla.server.errors import EntityAlreadyExistsError, EntityNotFoundError, ForbiddenOperationError, WrongTaskError
 from argilla.server.models import User, Workspace
 from argilla.server.policies import DatasetPolicy, DatasetSettingsPolicy, is_authorized
-from argilla.server.schemas.datasets import CreateDatasetRequest, Dataset
+from argilla.server.schemas.v0.datasets import CreateDatasetRequest, Dataset
 
 
 class ServiceBaseDataset(BaseDatasetDB):