+++ tests/functional_tests/search/test_search_service.py
@@ -25,7 +25,7 @@ from argilla.server.daos.backend import GenericElasticEngineBackend
 from argilla.server.daos.backend.search.query_builder import EsQueryBuilder
 from argilla.server.daos.models.datasets import BaseDatasetDB
 from argilla.server.daos.records import DatasetRecordsDAO
-from argilla.server.schemas.datasets import Dataset
+from argilla.server.schemas.v0.datasets import Dataset
 from argilla.server.services.metrics import MetricsService, ServicePythonMetric
 from argilla.server.services.search.model import ServiceSortConfig
 from argilla.server.services.search.service import SearchRecordsService