+++ tests/server/api/v1/test_questions.py
@@ -12,23 +12,25 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from datetime import datetime
 from typing import TYPE_CHECKING, Type
 from uuid import uuid4
 
 import pytest
 from argilla._constants import API_KEY_HEADER_NAME
-from argilla.server.models import DatasetStatus, Question
+from argilla.server.models import DatasetStatus, Question, UserRole
 from fastapi.testclient import TestClient
 from sqlalchemy import func, select
+from thinc.api import MultiSoftmax
 
 from tests.factories import (
     AnnotatorFactory,
     DatasetFactory,
     LabelSelectionQuestionFactory,
     MultiLabelSelectionQuestionFactory,
+    RankingQuestionFactory,
     RatingQuestionFactory,
     TextQuestionFactory,
+    UserFactory,
 )
 
 if TYPE_CHECKING:
@@ -38,29 +40,30 @@ if TYPE_CHECKING:
 
 
 @pytest.mark.parametrize(
-    "QuestionFactory, expected_settings",
+    "QuestionFactory, payload, expected_settings",
     [
-        (TextQuestionFactory, {"type": "text", "use_markdown": False}),
         (
-            RatingQuestionFactory,
+            TextQuestionFactory,
             {
-                "type": "rating",
-                "options": [
-                    {"value": 1},
-                    {"value": 2},
-                    {"value": 3},
-                    {"value": 4},
-                    {"value": 5},
-                    {"value": 6},
-                    {"value": 7},
-                    {"value": 8},
-                    {"value": 9},
-                    {"value": 10},
-                ],
+                "title": "New Title",
+                "description": "New Description",
+                "settings": {"type": "text", "use_markdown": True},
             },
+            {"type": "text", "use_markdown": True},
+        ),
+        (
+            TextQuestionFactory,
+            {"description": None},
+            {"type": "text", "use_markdown": False},
+        ),
+        (
+            TextQuestionFactory,
+            {"name": "New Name", "required": True, "dataset_id": str(uuid4())},
+            {"type": "text", "use_markdown": False},
         ),
         (
             LabelSelectionQuestionFactory,
+            {"settings": {"type": "label_selection", "visible_options": 3}},
             {
                 "type": "label_selection",
                 "options": [
@@ -68,10 +71,12 @@ if TYPE_CHECKING:
                     {"value": "option2", "text": "Option 2", "description": None},
                     {"value": "option3", "text": "Option 3", "description": None},
                 ],
+                "visible_options": 3,
             },
         ),
         (
             MultiLabelSelectionQuestionFactory,
+            {"settings": {"type": "multi_label_selection", "visible_options": 3}},
             {
                 "type": "multi_label_selection",
                 "options": [
@@ -79,21 +84,161 @@ if TYPE_CHECKING:
                     {"value": "option2", "text": "Option 2", "description": None},
                     {"value": "option3", "text": "Option 3", "description": None},
                 ],
+                "visible_options": 3,
+            },
+        ),
+        (
+            LabelSelectionQuestionFactory,
+            {"settings": {"type": "label_selection", "visible_options": None}},
+            {
+                "type": "label_selection",
+                "options": [
+                    {"value": "option1", "text": "Option 1", "description": None},
+                    {"value": "option2", "text": "Option 2", "description": None},
+                    {"value": "option3", "text": "Option 3", "description": None},
+                ],
+                "visible_options": None,
+            },
+        ),
+    ],
+)
+@pytest.mark.parametrize("role", [UserRole.owner])
+@pytest.mark.asyncio
+async def test_update_question(
+    client: TestClient,
+    db: "AsyncSession",
+    QuestionFactory: Type["QuestionFactoryType"],
+    payload: dict,
+    expected_settings: dict,
+    role: UserRole,
+):
+    question = await QuestionFactory.create()
+    user = await UserFactory.create(role=role, workspaces=[question.dataset.workspace])
+
+    response = client.patch(
+        f"/api/v1/questions/{question.id}", headers={API_KEY_HEADER_NAME: user.api_key}, json=payload
+    )
+
+    title = payload.get("title") or question.title
+    description = payload.get("description") or question.description
+
+    assert response.status_code == 200
+    assert response.json() == {
+        "id": str(question.id),
+        "name": question.name,
+        "title": title,
+        "description": description,
+        "required": False,
+        "settings": expected_settings,
+        "dataset_id": str(question.dataset_id),
+        "inserted_at": question.inserted_at.isoformat(),
+        "updated_at": question.updated_at.isoformat(),
+    }
+
+    question = await db.get(Question, question.id)
+    assert question.title == title
+    assert question.description == description
+    assert question.settings == expected_settings
+
+
+@pytest.mark.parametrize(
+    "QuestionFactory, payload",
+    [
+        (TextQuestionFactory, {"title": None, "description": None, "settings": None}),
+        (TextQuestionFactory, {"settings": {"type": "text"}}),
+        (TextQuestionFactory, {"settings": {"type": "text", "use_markdown": None}}),
+        (TextQuestionFactory, {"title": "New Title", "settings": {"type": "label_selection"}}),
+        (
+            RatingQuestionFactory,
+            {"settings": {"type": "rating", "options": [{"value": 94}, {"value": 95}, {"value": 96}]}},
+        ),
+        (LabelSelectionQuestionFactory, {"settings": {"type": "label_selection", "visible_options": -5}}),
+        (MultiLabelSelectionQuestionFactory, {"settings": {"type": "multi_label_selection", "visible_options": -5}}),
+        (
+            RankingQuestionFactory,
+            {
+                "settings": {
+                    "type": "ranking",
+                    "options": [
+                        {"value": "option-a", "text": "Option A", "description": None},
+                        {"value": "option-b", "text": "Option B", "description": None},
+                        {"value": "option-c", "text": "Option C", "description": None},
+                    ],
+                }
             },
         ),
     ],
 )
 @pytest.mark.asyncio
-async def test_delete_question(
-    client: TestClient,
-    db: "AsyncSession",
-    owner_auth_header: dict,
-    QuestionFactory: Type["QuestionFactoryType"],
-    expected_settings: dict,
+async def test_update_question_with_invalid_settings(
+    client: TestClient, owner_auth_header: dict, QuestionFactory: Type["QuestionFactoryType"], payload: dict
 ):
-    question = await QuestionFactory.create(name="name", title="title", description="description")
+    question = await QuestionFactory.create()
 
-    response = client.delete(f"/api/v1/questions/{question.id}", headers=owner_auth_header)
+    response = client.patch(f"/api/v1/questions/{question.id}", headers=owner_auth_header, json=payload)
+
+    assert response.status_code == 422
+
+
+@pytest.mark.asyncio
+async def test_update_question_with_invalid_payload(client: TestClient, owner_auth_header: dict):
+    question = await TextQuestionFactory.create()
+
+    response = client.patch(
+        f"/api/v1/questions/{question.id}",
+        headers=owner_auth_header,
+        json={"title": {"this": "is", "not": "valid"}, "settings": {"use_markdown": "no"}},
+    )
+
+    assert response.status_code == 422
+
+
+@pytest.mark.asyncio
+async def test_update_question_non_existent(client: TestClient, owner_auth_header: dict):
+    response = client.patch(
+        f"/api/v1/questions/{uuid4()}",
+        headers=owner_auth_header,
+        json={"title": "New Title", "settings": {"type": "text", "use_markdown": True}},
+    )
+
+    assert response.status_code == 404
+
+
+@pytest.mark.asyncio
+async def test_update_question_as_admin_from_different_workspace(client: TestClient):
+    question = await TextQuestionFactory.create()
+    user = await UserFactory.create(role=UserRole.admin)
+
+    response = client.patch(
+        f"/api/v1/questions/{question.id}",
+        headers={API_KEY_HEADER_NAME: user.api_key},
+        json={"title": "New Title", "settings": {"type": "text", "use_markdown": True}},
+    )
+
+    assert response.status_code == 403
+
+
+@pytest.mark.asyncio
+async def test_update_question_as_annotator(client: TestClient):
+    question = await TextQuestionFactory.create()
+    user = await AnnotatorFactory.create(workspaces=[question.dataset.workspace])
+
+    response = client.patch(
+        f"/api/v1/questions/{question.id}",
+        headers={API_KEY_HEADER_NAME: user.api_key},
+        json={"title": "New Title", "settings": {"type": "text", "use_markdown": True}},
+    )
+
+    assert response.status_code == 403
+
+
+@pytest.mark.parametrize("role", [UserRole.admin, UserRole.owner])
+@pytest.mark.asyncio
+async def test_delete_question(client: TestClient, db: "AsyncSession", role: UserRole):
+    question = await TextQuestionFactory.create(name="name", title="title", description="description")
+    user = await UserFactory.create(role=role, workspaces=[question.dataset.workspace])
+
+    response = client.delete(f"/api/v1/questions/{question.id}", headers={API_KEY_HEADER_NAME: user.api_key})
 
     assert response.status_code == 200
     assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
@@ -105,20 +250,21 @@ async def test_delete_question(
         "title": "title",
         "description": "description",
         "required": False,
-        "settings": expected_settings,
+        "settings": {"type": "text", "use_markdown": False},
         "dataset_id": str(question.dataset_id),
-        "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
-        "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
+        "inserted_at": question.inserted_at.isoformat(),
+        "updated_at": question.updated_at.isoformat(),
     }
 
 
 @pytest.mark.asyncio
-async def test_delete_question_without_authentication(client: TestClient, db: "AsyncSession"):
+async def test_delete_question_as_admin_from_different_workspace(client: TestClient, db: "AsyncSession"):
+    user = await UserFactory.create(role=UserRole.admin)
     question = await TextQuestionFactory.create()
 
-    response = client.delete(f"/api/v1/questions/{question.id}")
+    response = client.delete(f"/api/v1/questions/{question.id}", headers={API_KEY_HEADER_NAME: user.api_key})
 
-    assert response.status_code == 401
+    assert response.status_code == 403
     assert (await db.execute(select(func.count(Question.id)))).scalar() == 1
 
 
@@ -130,6 +276,16 @@ async def test_delete_question_as_annotator(client: TestClient, db: "AsyncSessio
     response = client.delete(f"/api/v1/questions/{question.id}", headers={API_KEY_HEADER_NAME: annotator.api_key})
 
     assert response.status_code == 403
+    assert (await db.execute(select(func.count(Question.id)))).scalar() == 1
+
+
+@pytest.mark.asyncio
+async def test_delete_question_without_authentication(client: TestClient, db: "AsyncSession"):
+    question = await TextQuestionFactory.create()
+
+    response = client.delete(f"/api/v1/questions/{question.id}")
+
+    assert response.status_code == 401
     assert (await db.execute(select(func.count(Question.id)))).scalar() == 1